Update landingSite.js
Created tile objects. Function loops through vertical and horizontal grid lines and pushes a tile object into an array for each tile on canvas. 

Implemented mouse controls. When user clicks on a tile, event listener determines which tile was clicked and if the player is able to move to that space. Players movement property is updated.