Update titleScreen.js
Updated start/continue button logic. If user presses "start new game" but already has a save file, prompt asks if user wants to delete save file and start a new game. If user clicks yes, localStorage is cleared and user is sent to the start of the game. Continue button only enters game is a save file is already present, otherwise an error message telling the user to start a new game instead is generated.