Update global.js
Updated enemy movement to prevent multiple enemies from landing on the player at the same time. If player is currently in a battle, other close by enemies will move their full range of spaces instead of landing on the player. Updated health bar draw function to make the width responsive to the health stat of each character as the active character changes.