Update global.js
Created ranged attack method for red nomad class. When enemy lands within a set number of tiles from the player, enemy attacks player without initiating a battle. The number of spaces away the enemy can attack from is determined by their range property. Began work on mini map. Smaller version of grid tiles are drawn on a specified area of the canvas with tiles representing a space that a building is on drawn a different color.